unify network and shared memory modes.
Only for pulseaudio clients in this patch. The aim is to deduplicate
code, so it's better to unify the output drivers as well in the future
instead of duplicating these changes to the alsa and raw variants.

The patch splits off the initialization and audio data transfer for both
receive methods and generalizes access to the audio data and stream
properties so that they're receiver encoding independent.

The main loop simply calls receiver_rcv_fn, a function pointer to a
specific receiver implementation, which fills the supplied receiver_data
structure with audio data and the stream properties. The rest of the main
loop can then control the output driver from this data.