Fix installer failure on different drives (#113)
Windows keeps track of the current working directory for each drive
separately.  With RMB -> Run as Admin, the script starts off in
`C:\Windows\system32`; if the installer lives on a different drive (e.g.
D:), we need to explicitly change to that drive as well as changing the
CWD on it.